Model: Working on creatures.

+++ Backend/Model/Battle/Stats.cs
@@ -2,9 +2,9 @@ namespace Model.Battle
 {
     struct Stats
     {
-        public int Attack { get; set; }
-        public int Defense { get; set; }
-        public int Speed { get; set; }
-        public int Health { get; set; }
+        public float Attack { get; set; }
+        public float Defense { get; set; }
+        public float Speed { get; set; }
+        public float Health { get; set; }
     }
 }
